add verbose and force command line options

+++ dban-install.sh
@@ -10,26 +10,30 @@ file=dban-2.3.0_i586.iso
 
 cd $workdir
 
-curl -z $file --location --silent -o $file $url
+ret="$(curl -z $file --location ${curlargs} --write-out '%{http_code}' -o $file $url)"
+if [ $ret = "200" -o -n "$force" ] ; then
+	echo "dban was updated"
+	mkdir -p dban-tmp
 
-mkdir dban-tmp
+	# we need to extract files from the dban iso
+	# not sure of a good cli command to do this :/
+	mount -o loop $file dban-tmp
 
-# we need to extract files from the dban iso
-# not sure of a good cli command to do this :/
-mount -o loop $file dban-tmp
 
-#ls -lh $file dban-tmp
-
-mkdir -p ${tftpdir}/dban
-cp dban-tmp/dban.bzi ${tftpdir}/dban
-umount dban-tmp
+	mkdir -p ${tftpdir}/dban
+	cp dban-tmp/dban.bzi ${tftpdir}/dban
+	umount dban-tmp
 
 cat <<ABC123
 # netboot setup for DBAN
 LABEL DBAN
+  MENU LABEL ^DBAN
   KERNEL dban/dban.bzi
+  APPEND nuke="dwipe"
 
 # XXX add an autonuke option
 ABC123
+
+fi
 
 # APPEND nuke="dwipe --autonuke" silent
